+++ wagtail_embed_videos/admin_urls.py
@@ -1,6 +1,5 @@
 from django.urls import path
 
-from wagtail_embed_videos.views import chooser
 from wagtail_embed_videos.views import embed_videos
 
 app_name = "wagtail_embed_videos"
@@ -12,16 +11,4 @@ urlpatterns = [
     path("<int:embed_video_id>/delete/", embed_videos.delete, name="delete"),
     path("add/", embed_videos.add, name="add"),
     path("usage/<int:embed_video_id>/", embed_videos.usage, name="embed_video_usage"),
-    path("chooser/", chooser.ChooseView.as_view(), name="chooser"),
-    path(
-        "chooser/results/",
-        chooser.ChooseResultsView.as_view(),
-        name="chooser_results",
-    ),
-    path(
-        "chooser/<int:embed_video_id>/",
-        chooser.embed_video_chosen,
-        name="embed_video_chosen",
-    ),
-    path("chooser/upload/", chooser.chooser_upload, name="chooser_upload"),
 ]