+++ wagtail_embed_videos/permissions.py
@@ -1,9 +1,7 @@
 from django.dispatch import receiver
 from django.test.signals import setting_changed
 
-from wagtail.core.permission_policies.collections import (
-    CollectionOwnershipPermissionPolicy,
-)
+from wagtail.permission_policies.collections import CollectionOwnershipPermissionPolicy
 
 from wagtail_embed_videos import get_embed_video_model
 from wagtail_embed_videos.models import EmbedVideo