+++ wagtail_embed_videos/views/embed_videos.py
@@ -16,8 +16,7 @@ from wagtail.admin.auth import PermissionPolicyChecker
 from wagtail.admin.forms.search import SearchForm
 from wagtail.admin.models import popular_tags_for_model
 from wagtail.admin.views.pages.utils import get_valid_next_url_from_request
-from wagtail.core.models import Collection
-from wagtail.search import index as search_index
+from wagtail.models import Collection
 
 from wagtail_embed_videos import get_embed_video_model
 from wagtail_embed_videos.forms import get_embed_video_form
@@ -25,27 +24,69 @@ from wagtail_embed_videos.permissions import permission_policy
 
 permission_checker = PermissionPolicyChecker(permission_policy)
 
-INDEX_PAGE_SIZE = getattr(settings, "WAGTAILEMBEDVIDEOS_INDEX_PAGE_SIZE", 20)
+INDEX_PAGE_SIZE = getattr(settings, "WAGTAILEMBEDVIDEOS_INDEX_PAGE_SIZE", 30)
 USAGE_PAGE_SIZE = getattr(settings, "WAGTAILEMBEDVIDEOS_USAGE_PAGE_SIZE", 20)
 
 
 class BaseListingView(TemplateView):
+    ENTRIES_PER_PAGE_CHOICES = sorted({10, 30, 60, 100, 250, INDEX_PAGE_SIZE})
+    ORDERING_OPTIONS = {
+        "-created_at": _("Newest"),
+        "created_at": _("Oldest"),
+    }
+    default_ordering = "-created_at"
+
     @method_decorator(permission_checker.require_any("add", "change", "delete"))
     def get(self, request):
         return super().get(request)
 
+    def get_num_entries_per_page(self):
+        entries_per_page = self.request.GET.get("entries_per_page", INDEX_PAGE_SIZE)
+        try:
+            entries_per_page = int(entries_per_page)
+        except ValueError:
+            entries_per_page = INDEX_PAGE_SIZE
+        if entries_per_page not in self.ENTRIES_PER_PAGE_CHOICES:
+            entries_per_page = INDEX_PAGE_SIZE
+
+        return entries_per_page
+
+    def get_valid_orderings(self):
+        return self.ORDERING_OPTIONS
+
+    def get_ordering(self):
+        # TODO: remove this method when this view will be based on the
+        # generic model index view from wagtail.admin.views.generic.models.IndexView
+        ordering = self.request.GET.get("ordering")
+        if ordering is None or ordering not in self.get_valid_orderings():
+            ordering = self.default_ordering
+        return ordering
+
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
 
-        # Get embed videos (filtered by user permission)
+        # Get ordering
+        ordering = self.get_ordering()
+
+        # Get embed videos (filtered by user permission and ordered by `ordering`)
         embed_videos = (
             permission_policy.instances_user_has_any_permission_for(
                 self.request.user,
                 ["change", "delete"],
             )
-            .order_by("-created_at")
+            .order_by(ordering)
             .select_related("collection")
         )
+
+        # Filter by collection
+        self.current_collection = None
+        collection_id = self.request.GET.get("collection_id")
+        if collection_id:
+            try:
+                self.current_collection = Collection.objects.get(id=collection_id)
+                embed_videos = embed_videos.filter(collection=self.current_collection)
+            except (ValueError, Collection.DoesNotExist):
+                pass
 
         # Search
         query_string = None
@@ -61,16 +102,6 @@ class BaseListingView(TemplateView):
         else:
             self.form = SearchForm(placeholder=_("Search embed videos"))
 
-        # Filter by collection
-        self.current_collection = None
-        collection_id = self.request.GET.get("collection_id")
-        if collection_id:
-            try:
-                self.current_collection = Collection.objects.get(id=collection_id)
-                embed_videos = embed_videos.filter(collection=self.current_collection)
-            except (ValueError, Collection.DoesNotExist):
-                pass
-
         # Filter by tag
         self.current_tag = self.request.GET.get("tag")
         if self.current_tag:
@@ -79,7 +110,8 @@ class BaseListingView(TemplateView):
             except AttributeError:
                 self.current_tag = None
 
-        paginator = Paginator(embed_videos, per_page=INDEX_PAGE_SIZE)
+        entries_per_page = self.get_num_entries_per_page()
+        paginator = Paginator(embed_videos, per_page=entries_per_page)
         embed_videos = paginator.get_page(self.request.GET.get("p"))
 
         next_url = reverse("wagtail_embed_videos:index")
@@ -93,6 +125,10 @@ class BaseListingView(TemplateView):
                 "query_string": query_string,
                 "is_searching": bool(query_string),
                 "next": next_url,
+                "entries_per_page": entries_per_page,
+                "ENTRIES_PER_PAGE_CHOICES": self.ENTRIES_PER_PAGE_CHOICES,
+                "current_ordering": ordering,
+                "ORDERING_OPTIONS": self.ORDERING_OPTIONS,
             },
         )
 
@@ -161,9 +197,6 @@ def edit(request, embed_video_id):
         )
         if form.is_valid():
             form.save()
-
-            # Reindex the embed video to make sure all tags are indexed
-            search_index.insert_or_update_object(embed_video)
 
             edit_url = reverse("wagtail_embed_videos:edit", args=(embed_video.id,))
             redirect_url = "wagtail_embed_videos:index"
@@ -244,9 +277,6 @@ def add(request):
         )
         if form.is_valid():
             form.save()
-
-            # Reindex the embed video to make sure all tags are indexed
-            search_index.insert_or_update_object(embed_video)
 
             messages.success(
                 request,